Make --peers handling match cli help and ETCDCTL_PEERS handling
The cli help for etcdctl says that you can supply "a
comma-delimited list of machine addresses in the cluster". This
does in fact work for the ETCDCTL_PEERS environment variable,
but does not work if you use the --peers flag.

Change the handling of peer configuration so it works uniformly
between the --peers flag and the ETCDCTL_PEERS environment
variable.