support 'etcdctl mkdir -p'
The 'mkdir' subcommand is not idempotent, and it is awkward to do
'etcdctl ls /foo || etcdctl mkdir /foo'.

Add a 'p' flag to the mkdir command so that, when this flag is passed,
it exits with success without trying to create a dir that already
exists.

This allows its easy use in fixture style shell scripts that create a
whole etcdctl structure and run everytime a build is performed.